+++ src/tests/conftest.py
@@ -373,6 +373,46 @@ def mock_pipeline_no_task(mock_app, mock_text_in, mock_text_out):
     return p
 
 
+@pytest.fixture
+def mock_timestamped_partitioned_dir(tmp_path):
+    partitioned_dir = tmp_path / "timestamped_partitioned"
+    partitioned_dir.mkdir(parents=True, exist_ok=True)
+    (partitioned_dir / "part_00.txt").write_text("part 00")
+    (partitioned_dir / "part_01.txt").write_text("part 01")
+    return partitioned_dir
+
+
+@pytest.fixture
+def mock_example01(mock_app, mock_timestamped_partitioned_dir, mock_text_in):
+
+    inputs = [{"name": "text_in", "config": json.dumps(
+        {"type": "text.TextDataset", "filepath": str(mock_text_in)})}]
+    outputs = [{"name": "timestamped_partitioned", "config": json.dumps(
+        {"type": "partitions.PartitionedDataset",
+         "path": str(mock_timestamped_partitioned_dir),
+         "filename_suffix": ".txt",
+         "dataset": {"type": "text.TextDataset"}})}]
+    parameters = [{"name": "example", "value": "hello"}]
+    tags = [{"key": "author", "value": "harinlee83"}, {
+        "key": "package", "value": "kedro-graphql"}]
+
+    p = Pipeline(
+        name="example01",
+        data_catalog=[DataSet(**i) for i in inputs] + [DataSet(**o) for o in outputs],
+        parameters=[Parameter(**p) for p in parameters],
+        tags=[Tag(**p) for p in tags],
+        status=[PipelineStatus(state=State.STAGED,
+                               runner=mock_app.config["KEDRO_GRAPHQL_RUNNER"],
+                               session=mock_app.kedro_session.session_id,
+                               started_at=datetime.now(),
+                               task_name=str(run_pipeline))]
+    )
+
+    p.created_at = datetime.now()
+    p = mock_app.backend.create(p)
+    return p
+
+
 @pytest.fixture(autouse=True)
 def delete_pipeline_collection(mock_app):
     # Will be executed before the first test